turbo_az/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
distlib==0.3.8
Django==5.0.6
django-autocomplete-light==3.11.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-rest-framework==0.1.0
django-timezone-field==7.0
djangorestframework==3.15.2
dnspython==2.6.1
email_validator==2.2.0
Faker==25.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.13.4
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.3.7
langchain==0.2.5
langchain-core==0.2.9
langchain-text-splitters==0.2.1
langsmith==0.1.81
logging==0.4.9.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
MouseInfo==0.1.3
multidict==6.0.5
numpy==1.26.4
opencv-python==4.9.0.80
orjson==3.10.5
outcome==1.3.0.post0
packaging==24.1
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.0
prompt_toolkit==3.0.47
PyAutoGUI==0.9.54
pydantic==2.7.4
pydantic_core==2.18.4
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pyperclip==1.8.2
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
pytweening==1.2.0
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.8
requests==2.32.3
rich==13.7.1
rubicon-objc==0.4.7
selenium==4.20.0
setuptools==69.5.1
shell==1.0.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
sqlparse==0.5.0
starlette==0.37.2
tenacity==8.4.2
text-unidecode==1.3
trio==0.25.0
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
vine==5.1.0
virtualenv==20.25.2
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
